requirements.txt
cryptography==46.0.3
google_api_python_client==2.176.0
google_auth_oauthlib==1.2.2
protobuf==6.33.2
PyQt6==6.10.1
pyqt6_sip==13.10.2
speedtest_cli==2.1.3
